requirements.txt
aiohttp
pandas
colorama
pyfiglet
tqdm
geoip2
python-dotenv
openpyxl
